rename: 리마인더 알림 코드 분리 및 패키지 변경
