+++ presentation/src/main/java/com/hihihihi/gureumpage/notification/reminder/DailyReminderWorker.kt
@@ -1,4 +1,4 @@
-package com.hihihihi.gureumpage.notification
+package com.hihihihi.gureumpage.notification.reminder
 
 import android.Manifest
 import android.content.Context
@@ -9,17 +9,11 @@ import androidx.core.content.ContextCompat
 import androidx.core.net.toUri
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
-import androidx.work.ExistingWorkPolicy
-import androidx.work.OneTimeWorkRequestBuilder
-import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import com.hihihihi.gureumpage.notification.common.Channels
 import com.hihihihi.gureumpage.notification.common.NotificationFactory
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
-import java.time.Duration
-import java.time.ZonedDateTime
-import java.util.concurrent.TimeUnit
 
 @HiltWorker
 class DailyReminderWorker @AssistedInject constructor(
@@ -47,23 +41,5 @@ class DailyReminderWorker @AssistedInject constructor(
             factory.notify("reminder:daily", 10001, notification)
         }
         return Result.success()
-    }
-}
-
-object ReminderScheduler {
-    fun scheduleDaily(context: Context, hour: Int = 20, minute: Int = 0) {
-        val now = ZonedDateTime.now()
-        var next = now.withHour(hour).withMinute(minute).withSecond(0).withNano(0)
-        if (!next.isAfter(now)) next = next.plusDays(1)
-        val delayMin = Duration.between(now, next).toMinutes()
-
-        val request = OneTimeWorkRequestBuilder<DailyReminderWorker>()
-            .setInitialDelay(delayMin, TimeUnit.MINUTES)
-            .addTag("daily-reminder-once")
-            .build()
-
-        WorkManager.getInstance(context).enqueueUniqueWork(
-            "daily-reminder", ExistingWorkPolicy.REPLACE, request
-        )
     }
 }
